tests: getting all tests to pass again
